+++ src/Number.h
@@ -5,4 +5,4 @@
 #define Number_BufferSize 15
 #endif
 
-String Number_Format(float number, size_t decimals);
+String Number_Format(s32 number, size_t decimals);
